Refactor and update multiple intro classes

+++ kodilla-intro/src/main/java/com/kodilla/AdvCalculator.java
@@ -1,5 +1,4 @@
 package com.kodilla;
-
 public class AdvCalculator {
     public double calculate() {
         String userSelected = UserDialogs.getUserSelection();
@@ -8,7 +7,7 @@ public class AdvCalculator {
         return switch (userSelected) {
             case "ADD" -> a + b;
             case "SUB" -> a - b;
-            case "DIV" -> a / (double)b; // dzielenie z rzutowaniem na double
+            case "DIV" -> a / b;
             case "MUL" -> a * b;
             default -> 0;
         };
